IOT_SMART_HOME-main/requirements.txt
paho-mqtt==1.6.1
streamlit==1.23.1
pydantic==1.10.8 
black==23.3.0 
pytest==7.3.1
asyncio==3.4.3 
icecream==2.1.3
streamlit-autorefresh==0.0.1

sqlite-web==0.2.1